Consolidate demo data, load sample data at startup
