+++ api/src/main/resources/application.conf
@@ -13,6 +13,8 @@ hmda {
     port = 8080
     port = ${?HMDA_HTTP_PORT}
   }
+  isDemo = true
+  isDemo = ${?HMDA_IS_DEMO}
 }
 
 akka.persistence.journal.plugin = "akka.persistence.journal.leveldb"
